With this being a module we can now call functions instead of including whole file
